cli: refactor detach to the new approach
